requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-dotenv
python-multipart
pydantic
pytest
httpx
ocpp==0.7.2
jinja2
